pr review, push forgotten changes

+++ include/maplab-realsense/zr300-config.h
@@ -1,6 +1,8 @@
 #ifndef MAPLAB_REALSENSE_ZR300_CONFIG_H_
 #define MAPLAB_REALSENSE_ZR300_CONFIG_H_
 
+#include <string>
+
 #include <librealsense/rs.hpp>
 #include <librealsense/rsutil.h>
 #include <ros/ros.h>
@@ -57,7 +59,7 @@ struct ZR300Config {
   bool infrared_enabled = true;
   int infrared_subsample_factor = 1;
   bool infrared_auto_exposure = true;
-  int infrared_gain = 0.5;
+  float infrared_gain = 0.5;
   int infrared_exposure = 0;
   bool infrared_emitter_enabled = true;
